Remove the strava service
