+++ Guppi.Application/Queries/Strava/AuthorizationRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace Guppi.Infrastructure.Services.Strava
+namespace Guppi.Application.Queries.Strava
 {
 #pragma warning disable IDE1006 // Naming Styles
     public class AuthorizationRequest
