+++ Guppi.Domain/Interfaces/IStravaService.cs
@@ -6,7 +6,5 @@ namespace Guppi.Domain.Interfaces
     public interface IStravaService
     {
         Task<IEnumerable<Domain.Entities.Strava.StravaActivity>> GetActivities();
-
-        void Configure();
     }
 }
